high-risk-project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
numpy==1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
openai
streamlit
boto3
fpdf